fix(dpp)!: version-gate the byte array encoding-stability check (protocol 12)
Address review: a consensus tightening must not change behavior for already
released protocol versions. validate_update is dispatched on a feature version
that protocol versions 1-11 pin to 0, so adding the check in validate_update_v0
would alter consensus for deployed versions and could split upgraded vs
unupgraded validators during a rolling rollout.

- move the check into a new validate_update_v1 (= v0 checks + byte array
  encoding stability); validate_update_v0 is restored bit-exact
- dispatcher: add arm 1 => validate_update_v1
- DPP_VALIDATION_VERSIONS_V3.document_type.validate_update: 0 -> 1 (V3 is used
  only by protocol 12 / release 3.1.0), so the rule activates atomically at that
  protocol upgrade rather than on binary rollout
- tests now exercise the public validate_update dispatcher (latest selects v1)
  plus a pre-activation version (protocol 11 still accepts the flip), so a
  missing arm or stale feature-version constant fails a test
- e2e: assert the decode-failure error variant instead of substring matching
